Simplify the way to populate edit pdf parameters
