+++ samples/EditAdvanced.cs
@@ -25,18 +25,19 @@ namespace Samples
             var imageFile = task.AddFile("your_image.jpg");
             var svgFile = task.AddFile("your_image.svg");
 
-            // Create edit task params
-            var editParams = EditParams.New();
+            // Create edit task params builder
+            var builder = new EditParamBuilder();
 
             // Add elements to Editpdf task in order of drawing (important if elements overlap!)
+
             // Create ImageElement
-            var imageElement = editParams.AddImage(imageFile.ServerFileName);
+            var imageElement = builder.AddImage(imageFile.ServerFileName);
             imageElement.Coordinates = new Coordinate(300, 600);
             imageElement.Pages = 3;
             imageElement.Opacity = 40;
 
             // Create TextElement
-            var textElement = editParams.AddText("This is a sample text");
+            var textElement = builder.AddText("This is a sample text");
             textElement.Coordinates = new Coordinate(300, 600);
             textElement.Pages = 2;
             textElement.Align = TextAligments.Center;
@@ -45,13 +46,12 @@ namespace Samples
             textElement.FontStyle = TextFontStyles.Bold;
 
             // Create SvgElement
-            var svgElement = new SvgElement(svgFile.ServerFileName) 
-            {
-                Coordinates = new Coordinate(300, 300)
-            };
-            editParams.AddElement(svgElement);
+            var svgElement = builder.AddSvg(svgFile.ServerFileName);
+            svgElement.Coordinates = new Coordinate(300, 300);
 
-            //
+            // Create edit task params
+            var editParams = new EditParams(builder);
+             
             editParams.OutputFileName = "editpdf-advanced";
             
             // Proces added files,
